cleaner version

+++ Variable.java
@@ -9,6 +9,8 @@ public class Variable {
 
     protected int max;
     protected int min;
+
+    // following includes values [min-, min, min+, max-, max, max+] depending on the type of testing performed with the variable
     protected ArrayList<Integer> test_values;        // does not include the nominal as it functions separately in the logic
 
     // constructor
@@ -41,7 +43,7 @@ public class Variable {
     }
 
     public int nominal () {
-        return (this.max - this.min) / 2;       // DOUBT: what if this is a decimal?
+        return (int) (this.max - this.min) / 2;       // DOUBT: what if this is a decimal?
     }
 
     public int max_minus () {
